requirements.txt
discord.py
typing_extensions
python-dotenv